% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerator.R
\name{accelerator}
\alias{accelerator}
\title{Create an accelerator}
\usage{
accelerator(device_placement = TRUE, cpu = FALSE)
}
\arguments{
\item{device_placement}{(logical) whether the \code{accelerator} object should
handle device placement. Default: \code{TRUE}}

\item{cpu}{(logical) whether the training procedure should run on the CPU.}
}
\description{
Create an accelerator
}
