% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataloader.R
\name{as_dataloader}
\alias{as_dataloader}
\title{Creates a dataloader from its input}
\usage{
as_dataloader(x, ...)
}
\arguments{
\item{x}{the input object.}

\item{...}{used by the specialized methods.}
}
\description{
\code{as_dataloader} is used internally by luz to convert input
\code{data} and \code{valid_data} as passed to \link{fit} to a \link[torch:dataloader]{torch::dataloader}
}
\details{
\code{as_dataloader} methods should have sensible defaults for batch_size,
parallel workers, etc.

It allows users to quickly experiment with \link{fit} by not requiring
to create a \link[torch:dataset]{torch::dataset} and a \link[torch:dataloader]{torch::dataloader} in simple
experiments.
}
