% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{get_metrics}
\alias{get_metrics}
\alias{get_metrics.luz_module_fitted}
\title{Get metrics from the object}
\usage{
get_metrics(object, ...)

\method{get_metrics}{luz_module_fitted}(object, ...)
}
\arguments{
\item{object}{The object to query for metrics.}

\item{...}{Currently unused.}
}
\value{
A data.frame containing the metric values.
}
\description{
Get metrics from the object
}
\section{Methods (by class)}{
\itemize{
\item \code{get_metrics(luz_module_fitted)}: Extract metrics from a luz fitted model.

}}
