% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr-finder.R
\name{lr_finder}
\alias{lr_finder}
\title{Learning Rate Finder}
\usage{
lr_finder(
  object,
  data,
  steps = 100,
  start_lr = 0.1,
  end_lr = 1e-07,
  plot = TRUE
)
}
\arguments{
\item{object}{An nn_module that has been setup().}

\item{data}{(dataloader) A dataloader created with torch::dataloader()  used for learning rate finding.}

\item{steps}{(integer) The number of steps to iterate over in the learning rate finder. Default: 100.}

\item{start_lr}{(float) The largest learning rate. Default: 1e-1.}

\item{end_lr}{(float) The lowest learning rate. Default: 1e-7.}

\item{plot}{(bool) A logical indicator whether to print a plot of learnign rate vs. loss.}
}
\value{
A dataframe with two columns: learning rate and loss
}
\description{
Learning Rate Finder
}
