% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-resume.R
\name{luz_callback_auto_resume}
\alias{luz_callback_auto_resume}
\title{Resume training callback}
\usage{
luz_callback_auto_resume(path = "./state")
}
\arguments{
\item{path}{Path to save state files for the model.}
}
\description{
This callback allows you to resume training a model from a specified
directory.
}
\details{
When using it, model weights, optimizer state and learning rate schedulers
state is serialized each epoch. If something fails during training simply
re-running the same script will restart the model training from where it
stopped.
}
