% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{luz_callback_csv_logger}
\alias{luz_callback_csv_logger}
\title{CSV logger callback}
\usage{
luz_callback_csv_logger(path)
}
\arguments{
\item{path}{path to a file on disk.}
}
\description{
Logs metrics obtained during training a fiel on disk.
The file will have 1 line for each epoch/validation.
}
\seealso{
Other luz_callbacks: 
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_interrupt}()},
\code{\link{luz_callback_lr_scheduler}()},
\code{\link{luz_callback_metrics}()},
\code{\link{luz_callback_model_checkpoint}()},
\code{\link{luz_callback_profile}()},
\code{\link{luz_callback_progress}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
