% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{luz_callback_metrics}
\alias{luz_callback_metrics}
\title{Metrics callback}
\usage{
luz_callback_metrics()
}
\description{
Tracks metrics passed to \code{\link[=setup]{setup()}} during training and validation.
}
\details{
This callback takes care of 3 \code{ctx} attributes:
\itemize{
\item \code{ctx$metrics}: stores the metrics objects that are initialized once for epoch,
and are further \code{update()}d and \code{compute()}d every batch.
\item \code{ctx$losses}: stores the loss values in a named list with names ("train", "valid").
Each one is a sequence of the loss values per epoch.
\item \code{ctx$records$metrics}: Stores metrics per training/validation and epoch. The
structure is very similar to \code{ctx$losses}.
}
}
\note{
In general you won't need to explicitly use the metrics callback as it's
used by default in \code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.
}
\seealso{
Other luz_callbacks: 
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_progress}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
