% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-profile.R
\name{luz_callback_profile}
\alias{luz_callback_profile}
\title{Profile callback}
\usage{
luz_callback_profile()
}
\value{
A \code{luz_callback}
}
\description{
Computes the times for high-level operations in the training loops.
}
\details{
Records are saved in \code{ctx$records$profile}. Times are stored as seconds.
Data is stored in the following structure:
\itemize{
\item \strong{fit} time for the entire fit procedure.
\item \strong{epoch} times per epoch
\item \strong{(train/valid)_batch} time per batch of data processed, including data
acquisition and step.
\item \strong{(train/valid)_step} time per step (training or validation step) - only
the model step. (not including data acquisition and preprocessing)
}
}
\note{
In general you don't need to use these callback by yourself because it's always
included by default in \code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.
}
\examples{
profile_callback <- luz_callback_profile()

}
\seealso{
Other luz_callbacks: 
\code{\link{luz_callback_csv_logger}()},
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_interrupt}()},
\code{\link{luz_callback_lr_scheduler}()},
\code{\link{luz_callback_metrics}()},
\code{\link{luz_callback_mixup}()},
\code{\link{luz_callback_model_checkpoint}()},
\code{\link{luz_callback_progress}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback_use_best_model}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
