% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-resume.R
\name{luz_callback_resume_from_checkpoint}
\alias{luz_callback_resume_from_checkpoint}
\title{Allow resume model training from a specific checkpoint}
\usage{
luz_callback_resume_from_checkpoint(
  path,
  ...,
  restore_model_state = TRUE,
  restore_records = FALSE,
  restore_optimizer_state = FALSE,
  restore_callbacks_state = FALSE
)
}
\arguments{
\item{path}{Path to the checkpoint that you want to resume.}

\item{...}{currently unused.}

\item{restore_model_state}{Wether to restore the model state from the callback.}

\item{restore_records}{Wether to restore records from the checkpoint.}

\item{restore_optimizer_state}{Wether to restore the optimizer state from the
checkpoint.}

\item{restore_callbacks_state}{Wether to restore the callbacks state from the
checkpoint.}
}
\description{
Allow resume model training from a specific checkpoint
}
\note{
Read the checkpointing article in the pkgdown website for more
information.
}
\seealso{
\code{\link[=luz_callback_model_checkpoint]{luz_callback_model_checkpoint()}}

Other luz_callbacks: 
\code{\link{luz_callback_auto_resume}()},
\code{\link{luz_callback_csv_logger}()},
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_interrupt}()},
\code{\link{luz_callback_keep_best_model}()},
\code{\link{luz_callback_lr_scheduler}()},
\code{\link{luz_callback_metrics}()},
\code{\link{luz_callback_mixed_precision}()},
\code{\link{luz_callback_mixup}()},
\code{\link{luz_callback_model_checkpoint}()},
\code{\link{luz_callback_profile}()},
\code{\link{luz_callback_progress}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
