% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback-validation-check.R
\name{luz_callback_validation_check}
\alias{luz_callback_validation_check}
\title{Validation Check}
\usage{
luz_callback_validation_check(batches = 2)
}
\arguments{
\item{batches}{Number of validation batches to check. Default is 2.}
}
\value{
A \code{luz_callback}.
}
\description{
Check validation loop before fitting model.
}
\note{
Usually the training loop is much longer than the validation
loop and issues with the validation loop aren't encountered until after
a long training runtime. This callback runs the validation loop first on
\code{batches} number of batches and then proceeds onto the standard
training process.

Printing can be disabled by passing \code{verbose = FALSE} to
\code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.
}
\seealso{
Other luz_callbacks: 
\code{\link{luz_callback_csv_logger}()},
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_interrupt}()},
\code{\link{luz_callback_lr_scheduler}()},
\code{\link{luz_callback_metrics}()},
\code{\link{luz_callback_model_checkpoint}()},
\code{\link{luz_callback_profile}()},
\code{\link{luz_callback_progress}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
