% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{luz_load}
\alias{luz_load}
\title{Load trained model}
\usage{
luz_load(path)
}
\arguments{
\item{path}{path in file system so save the object.}
}
\description{
Loads a fitted model. See documentation in \code{\link[=luz_save]{luz_save()}}.
}
\seealso{
Other luz_save: 
\code{\link{luz_save}()}
}
\concept{luz_save}
