% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{luz_load_checkpoint}
\alias{luz_load_checkpoint}
\title{Loads a checkpoint}
\usage{
luz_load_checkpoint(obj, path, ...)
}
\arguments{
\item{obj}{Object to which we want to laod the checkpoint.}

\item{path}{Path of the checkpoint on disk.}

\item{...}{unused. Is there to allow future extensions.}
}
\description{
Works with checkpoints created typically with \code{\link[=luz_callback_model_checkpoint]{luz_callback_model_checkpoint()}}.
}
