% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{luz_load_model_weights}
\alias{luz_load_model_weights}
\alias{luz_save_model_weights}
\title{Loads model weights into a fitted object.}
\usage{
luz_load_model_weights(obj, path, ...)

luz_save_model_weights(obj, path)
}
\arguments{
\item{obj}{luz object to which you want to copy the new weights.}

\item{path}{path to saved model in disk.}

\item{...}{other arguments passed to \code{\link[=torch_load]{torch_load()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
This can be useful when you have saved model checkpoints during training and
want to reload the best checkpoint in the end.
}
\section{Warning}{

\code{luz_save_model_weights} operates inplace, ie modifies the model object to contain the
new weights.
}

