% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{luz_metric_mae}
\alias{luz_metric_mae}
\title{Mean absolute error}
\usage{
luz_metric_mae()
}
\value{
Returns new luz metric.
}
\description{
Computes the mean absolute error.
}
\examples{
if (torch::torch_is_installed()) {
library(torch)
metric <- luz_metric_mae()
metric <- metric$new()
metric$update(torch_randn(100), torch_randn(100))
metric$compute()
}
}
\seealso{
Other luz_metrics: 
\code{\link{luz_metric_accuracy}()},
\code{\link{luz_metric_binary_accuracy_with_logits}()},
\code{\link{luz_metric_binary_accuracy}()},
\code{\link{luz_metric_binary_auroc}()},
\code{\link{luz_metric_mse}()},
\code{\link{luz_metric_multiclass_auroc}()},
\code{\link{luz_metric_rmse}()},
\code{\link{luz_metric}()}
}
\concept{luz_metrics}
