% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{luz_metric_rmse}
\alias{luz_metric_rmse}
\title{Root mean squared error}
\usage{
luz_metric_rmse()
}
\value{
Returns new luz metric.
}
\description{
Computes the root mean squared error.
}
\seealso{
Other luz_metrics: 
\code{\link{luz_metric_accuracy}()},
\code{\link{luz_metric_binary_accuracy_with_logits}()},
\code{\link{luz_metric_binary_accuracy}()},
\code{\link{luz_metric_binary_auroc}()},
\code{\link{luz_metric_mae}()},
\code{\link{luz_metric_mse}()},
\code{\link{luz_metric_multiclass_auroc}()},
\code{\link{luz_metric}()}
}
\concept{luz_metrics}
