% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{luz_metric_set}
\alias{luz_metric_set}
\title{Creates a metric set}
\usage{
luz_metric_set(metrics = NULL, train_metrics = NULL, valid_metrics = NULL)
}
\arguments{
\item{metrics}{A list of luz_metrics that are meant to be used in both training
and validation.}

\item{train_metrics}{A list of luz_metrics that are only used during training.}

\item{valid_metrics}{A list of luz_metrics that are only sued for validation.}
}
\description{
A metric set can be used to specify metrics that are only evaluated during
training, validation or both.
}
