% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{luz_metric_yardstick}
\alias{luz_metric_yardstick}
\title{Computes the average for any yardstick metric}
\usage{
luz_metric_yardstick(metric_nm, transform = NULL, ...)
}
\arguments{
\item{metric_nm}{Name of the metric from yardstick (without the \verb{_vec}).
For example \code{'accuracy'}, \code{'mae'}, etc.}

\item{transform}{A function of \code{preds} and \code{targets} that will be applied
to the values before computing the metric. This function is called after
moving \code{preds} and \code{targets} to R vectors.}

\item{...}{Additional parameters forwarded to the metric implementation in
yardstick.}
}
\value{
A luz metric object.
}
\description{
Allows using any yardstick metric with luz.
}
\section{Warning}{

The only transformation we do on the predicted values and in the
moving to R with \code{\link[torch:as_array]{torch::as_array()}}. However, many metrics in yardstick
expect that values are factors, or in other formats. In that case you can use
the \code{transform} argument to specify a transformation.
}

\examples{
if (torch::torch_is_installed()) {
x <- torch::torch_randn(100)
y <- torch::torch_randn(100)

m <- luz_metric_yardstick("mae")
m <- m$new()

m$update(x, y)
o <- m$compute()
}
}
