% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{luz_save}
\alias{luz_save}
\title{Saves luz objects to disk}
\usage{
luz_save(obj, path, ...)
}
\arguments{
\item{obj}{an object of class 'luz_module_fitted' as returned by
\code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.}

\item{path}{path in file system so save the object.}

\item{...}{currently unused.}
}
\description{
ALlows saving luz fitted models to the disk.
}
\seealso{
Other luz_save: 
\code{\link{luz_load}()}
}
\concept{luz_save}
