% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{nn_mixup_loss}
\alias{nn_mixup_loss}
\title{Loss to be used with \code{callbacks_mixup()}.}
\usage{
nn_mixup_loss(loss)
}
\arguments{
\item{loss}{the underlying loss module to call}
}
\description{
In the training phase, computes individual losses with regard to two targets, weights them item-wise,
and averages the linear combinations to yield the mean batch loss.
For validation and testing, defers to the passed-in loss.
}
