% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-mixup.R
\name{nnf_mixup}
\alias{nnf_mixup}
\title{Mixup logic}
\usage{
nnf_mixup(x, y, shuffle, weight)
}
\arguments{
\item{x}{an input batch}

\item{y}{a target batch}

\item{shuffle}{indices to be used to draw items to complement the "real batch"}

\item{weight}{weighting coefficient to be used by \code{torch_lerp()}}
}
\value{
A \code{list} of:
\itemize{
\item the new input batch, consisting of mixed \code{xs}
\item a \code{list} of:
\itemize{
\item a \code{list} of:
\itemize{
\item the original target \code{y1}
\item the mixed-in target \code{y2}
}
\item the mixing weights
}
}
}
\description{
Logic underlying \code{\link[=luz_callback_mixup]{luz_callback_mixup()}}.
}
\examples{
batch_x <- torch::torch_randn(c(10, 768))
batch_y <- torch::torch_randn(10)
shuffle <- torch::torch_tensor(10:1)
weight <- torch::torch_tensor(rep(0.9, 10))$view(c(10, 1))
nnf_mixup(batch_x, batch_y, shuffle, weight)

}
\seealso{
\code{\link[=luz_callback_mixup]{luz_callback_mixup()}}
}
