% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{set_hparams}
\alias{set_hparams}
\title{Set hyper-parameter of a module}
\usage{
set_hparams(module, ...)
}
\arguments{
\item{module}{An \code{nn_module} that has been \code{\link[=setup]{setup()}}.}

\item{...}{The parameters set here will be used to initialize the \code{nn_module}, ie they
are passed unchanged to the \code{initialize} method of the base \code{nn_module}.}
}
\value{
The same luz module
}
\description{
This function is used to define hyper-parameters before calling \code{fit} for
\code{luz_modules}.
}
\seealso{
Other set_hparam: 
\code{\link{set_opt_hparams}()}
}
\concept{set_hparam}
