% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{set_opt_hparams}
\alias{set_opt_hparams}
\title{Set optimizer hyper-parameters}
\usage{
set_opt_hparams(module, ...)
}
\arguments{
\item{module}{An \code{nn_module} that has been \code{\link[=setup]{setup()}}.}

\item{...}{The parameters passed here will be used to initialize the optimizers.
For example, if your optimizer is \code{optim_adam} and you pass \code{lr=0.1}, then the
\code{optim_adam} function is called with \code{optim_adam(parameters, lr=0.1)} when fitting
the model.}
}
\value{
The same luz module
}
\description{
This function is used to define hyper-parameters for the optimizer initialization
method.
}
\seealso{
Other set_hparam: 
\code{\link{set_hparams}()}
}
\concept{set_hparam}
