% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{evaluate}
\alias{evaluate}
\title{Evaluates a fitted model on a dataset}
\usage{
evaluate(
  object,
  data,
  ...,
  callbacks = list(),
  accelerator = NULL,
  verbose = NULL,
  dataloader_options = NULL
)
}
\description{
Evaluates a fitted model on a dataset
}
