% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{luz_callback_progress}
\alias{luz_callback_progress}
\title{Progress callback}
\usage{
luz_callback_progress()
}
\value{
A \code{luz_callback}
}
\description{
Responsible for printing progress during training.
}
\note{
In general you don't need to use these callback by yourself because it's always
included by default in \code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.

Printing can be disabled by passing \code{verbose=FALSE} to \code{\link[=fit.luz_module_generator]{fit.luz_module_generator()}}.
}
\seealso{
Other luz_callbacks: 
\code{\link{luz_callback_auto_resume}()},
\code{\link{luz_callback_csv_logger}()},
\code{\link{luz_callback_early_stopping}()},
\code{\link{luz_callback_interrupt}()},
\code{\link{luz_callback_keep_best_model}()},
\code{\link{luz_callback_lr_scheduler}()},
\code{\link{luz_callback_metrics}()},
\code{\link{luz_callback_mixup}()},
\code{\link{luz_callback_model_checkpoint}()},
\code{\link{luz_callback_profile}()},
\code{\link{luz_callback_resume_from_checkpoint}()},
\code{\link{luz_callback_train_valid}()},
\code{\link{luz_callback}()}
}
\concept{luz_callbacks}
